refine oneflow_compile api
