+++ src/onediff/infer_compiler/__init__.py
@@ -1,6 +1,7 @@
 import os
 import torch
 import oneflow as flow
+from .with_onef_graph import oneflow_compile 
 from .with_fx_interpreter import OneFlowInterpreter
 from .with_fx_graph import fx_node_tranform
